requirements.txt
# Core Dependencies
python-telegram-bot==20.3  # Official Telegram Bot API wrapper
httpx==0.25.0              # Async HTTP client for API requests

# Development Tools (Optional but recommended)
python-dotenv==1.0.0       # For local .env file loading
types-requests==2.31.0.1   # Type hints for better IDE support